Corrections for a few minor issues found via Eastwood lint tool
Type tags like ^boolean on Vars are ignored by the Clojure compiler.
To have any effect, they should be of the form ^{:tag 'boolean}.

Type tags on Java constructor expressions like ^Event (Event. payload)
are ignored by the Clojure compiler, and I think the type is
automatically inferred by the Clojure compiler anyway.

Added an import of a Java class that was used as a type tag.